+++ src/main/kotlin/com/loadingbyte/cinecred/project/StyleConstraints.kt
@@ -200,6 +200,8 @@ private val LETTER_STYLE_CONSTRAINTS: List<StyleConstraint<LetterStyle, *>> = li
 private const val BLUR_RADIUS_LIMIT = 200
 
 private val LAYER_CONSTRAINTS: List<StyleConstraint<Layer, *>> = listOf(
+    DoubleConstr(ERROR, Layer::gradientAngleDeg.st(), min = -90.0, max = 90.0),
+    DoubleConstr(ERROR, Layer::gradientExtentRfh.st(), min = 0.0),
     DoubleConstr(ERROR, Layer::stripeHeightRfh.st(), min = 0.0, minInclusive = false),
     DoubleConstr(ERROR, Layer::stripeCornerRadiusRfh.st(), min = 0.0, minInclusive = false),
     DoubleConstr(ERROR, Layer::stripeDashPatternRfh.st(), min = 0.0, minInclusive = false),